requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pydantic==1.10.12
google-generativeai==0.3.0
PyPDF2==3.0.1
PyMuPDF==1.22.5
Pillow==9.5.0
python-dotenv==1.0.0
aiofiles==23.1.0
httpx==0.24.1
pytest==7.3.1
pytest-asyncio==0.21.0
